fix: restore proper deprecation for database_less configuration
The database_less configuration option was removed in v9.1.0 without proper
deprecation, causing deployments to fail when upgrading if they had this
configuration in their production.rb or initializers.

This change restores backward compatibility by:

- Adding a minimal Configuration class that captures the deprecated database_less option
- Showing specific deprecation warnings when database_less is used
- Making the option a no-op (always returns false) since it's no longer needed
- Providing clear migration guidance for users to remove their configuration

The deprecation timeline indicates removal in v10.0.0, giving users time to
migrate away from this obsolete configuration without breaking deployments.

Also includes improvements to HasClosureTree include order dependency and
README formatting fixes to use inline return value comments.

Fixes the missing deprecation that was introduced when migrating to Zeitwerk.
